Be lax in an otherwise strict environment

This temporarily removes the shims added by strict so that you can execute code that would otherwise not be allowed

Usage

lax(code)

Arguments

code

Code to execute.

Examples

lax({
  mtcars[, 1]

  sapply(1:10, runif)
})
#> [[1]]
#> [1] 0.08075014
#> 
#> [[2]]
#> [1] 0.8343330 0.6007609
#> 
#> [[3]]
#> [1] 0.157208442 0.007399441 0.466393497
#> 
#> [[4]]
#> [1] 0.4977774 0.2897672 0.7328820 0.7725215
#> 
#> [[5]]
#> [1] 0.87460066 0.17494063 0.03424133 0.32038573 0.40232824
#> 
#> [[6]]
#> [1] 0.19566983 0.40353812 0.06366146 0.38870131 0.97554784 0.28989230
#> 
#> [[7]]
#> [1] 0.67838043 0.73531960 0.19595673 0.98053967 0.74152153 0.05144628 0.53021246
#> 
#> [[8]]
#> [1] 0.69582388 0.68855600 0.03123033 0.22556253 0.30083081 0.63646561 0.47902455
#> [8] 0.43217126
#> 
#> [[9]]
#> [1] 0.70643384 0.94857658 0.18033877 0.21689988 0.68016292 0.49884561 0.64167935
#> [8] 0.66028435 0.09602416
#> 
#> [[10]]
#>  [1] 0.7656002 0.7696748 0.9907123 0.9705209 0.3891828 0.4611865 0.3152418
#>  [8] 0.1746759 0.5315735 0.4936370
#>